Export json storage methods into modules
